+++ triple/lib/src/memento_mixin.dart
@@ -11,18 +11,18 @@ class _MutableIndex {
 @immutable
 mixin MementoMixin<State extends Object, Error extends Object> on Store<Error, State> {
   final _history = <Triple<Error, State>>[];
-  final int _historyLimit = 32;
+  final int _historyLimit = 64;
   final _MutableIndex _mutableIndex = _MutableIndex();
   int get _historyIndex => _mutableIndex.value;
   set _historyIndex(int value) => _mutableIndex.value = value;
 
-  /// Total size of history state caches;
+  /// Total size of history state caches
   int get historyLength => _history.length;
 
-  ///Return [true] if you can undo
+  /// Return [true] if you can undo
   bool canUndo() => _history.isNotEmpty && _historyIndex > 0;
 
-  ///Return [true] if you can redo
+  /// Return [true] if you can redo
   bool canRedo() => (_historyIndex + 1) < _history.length || triple.state != lastState.state;
 
   void _addHistory(Triple<Error, State> observableCache) {
@@ -52,7 +52,6 @@ mixin MementoMixin<State extends Object, Error extends Object> on Store<Error, S
   }
 
   ///Undo the last state value.
-  @protected
   void undo() {
     if (canUndo()) {
       _historyIndex = _historyIndex > _history.length ? math.max(_history.length - 1, 0) : _historyIndex - 1;
@@ -62,7 +61,6 @@ mixin MementoMixin<State extends Object, Error extends Object> on Store<Error, S
   }
 
   ///redo the last state value.
-  @protected
   void redo() {
     if (_historyIndex + 1 < _history.length) {
       _historyIndex++;
@@ -73,5 +71,10 @@ mixin MementoMixin<State extends Object, Error extends Object> on Store<Error, S
       // ignore: invalid_use_of_visible_for_overriding_member
       propagate(lastState);
     }
+  }
+
+  void clearHistory(int position) {
+    _history.clear();
+    _historyIndex = 0;
   }
 }